backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# Vector Database
pymilvus>=2.3.0

# LlamaIndex Core
llama-index-core>=0.10.0
llama-index-readers-file>=0.1.4
llama-index-embeddings-huggingface>=0.2.0
llama-index-vector-stores-milvus>=0.1.6
llama-index-llms-gemini>=0.1.0

# LLM & Embeddings
google-generativeai>=0.3.0
sentence-transformers>=2.2.0

# Agentic Workflow
langgraph>=0.0.20
langchain>=0.1.0
langchain-community>=0.0.10

# Data Collection
duckduckgo-search>=4.1.0
beautifulsoup4>=4.12.3
requests>=2.31.0

# Text Processing
PyPDF2>=3.0.0
tqdm>=4.66.0
